LPS-112372 Rename com.liferay.info.fields to com.liferay.info.field
