+++ modules/apps/info/info-api/bnd.bnd
@@ -8,8 +8,8 @@ Export-Package:\
 	com.liferay.info.display.field,\
 	com.liferay.info.display.request.attributes.contributor,\
 	com.liferay.info.display.url.provider,\
-	com.liferay.info.fields,\
-	com.liferay.info.fields.type,\
+	com.liferay.info.field,\
+	com.liferay.info.field.type,\
 	com.liferay.info.item,\
 	com.liferay.info.item.fields.reader,\
 	com.liferay.info.item.provider,\
