+++ modules/apps/info/info-api/src/main/java/com/liferay/info/item/fields/reader/InfoItemFieldReader.java
@@ -14,7 +14,7 @@
 
 package com.liferay.info.item.fields.reader;
 
-import com.liferay.info.fields.InfoField;
+import com.liferay.info.field.InfoField;
 
 /**
  * @author Jürgen Kappler
