+++ modules/apps/info/info-api/src/main/resources/com/liferay/info/field/packageinfo
