+++ modules/apps/info/info-impl/src/main/java/com/liferay/info/internal/display/contributor/InfoDisplayContributorWrapper.java
@@ -18,14 +18,14 @@ import com.liferay.info.display.contributor.InfoDisplayContributor;
 import com.liferay.info.display.contributor.InfoDisplayField;
 import com.liferay.info.display.contributor.InfoDisplayObjectProvider;
 import com.liferay.info.display.contributor.field.InfoDisplayContributorFieldType;
-import com.liferay.info.fields.InfoField;
-import com.liferay.info.fields.InfoFieldValue;
-import com.liferay.info.fields.InfoForm;
-import com.liferay.info.fields.InfoFormValues;
-import com.liferay.info.fields.type.ImageInfoFieldType;
-import com.liferay.info.fields.type.InfoFieldType;
-import com.liferay.info.fields.type.TextInfoFieldType;
-import com.liferay.info.fields.type.URLInfoFieldType;
+import com.liferay.info.field.InfoField;
+import com.liferay.info.field.InfoFieldValue;
+import com.liferay.info.field.InfoForm;
+import com.liferay.info.field.InfoFormValues;
+import com.liferay.info.field.type.ImageInfoFieldType;
+import com.liferay.info.field.type.InfoFieldType;
+import com.liferay.info.field.type.TextInfoFieldType;
+import com.liferay.info.field.type.URLInfoFieldType;
 import com.liferay.info.item.NoSuchInfoItemException;
 import com.liferay.info.item.provider.InfoItemFormProvider;
 import com.liferay.info.item.provider.InfoItemProvider;
